requirements.txt
accelerate==1.12.0
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
authlib==1.6.8
azure-core==1.38.1
azure-identity==1.25.2
beautifulsoup4==4.14.3
cachetools==7.0.1
catsu==0.0.2
cerebras-cloud-sdk==1.67.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chonkie==1.5.5
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
durationpy==0.10
elastic-transport==9.2.1
elasticsearch==9.3.0
exceptiongroup==1.3.1
fastuuid==0.14.0
filelock==3.24.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.48.0
google-genai==1.63.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
groq==1.0.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
html2text==2025.4.15
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==1.4.1
idna==3.11
importlib-metadata==8.7.1
jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
librt==0.8.1
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
mmh3==5.2.0
mpmath==1.3.0
msal==1.34.0
multidict==6.7.1
multiprocess==0.70.18
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.20.1
openai==2.21.0
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
overrides==7.7.0
packaging==24.2
pandas==2.3.3
pathspec==1.0.4
pgvector==0.3.6
pillow==12.1.1
pinecone==8.0.1
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
portalocker==3.2.0
posthog==5.4.0
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
psycopg2-binary==2.9.11
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pymilvus==2.6.9
pymongo==4.16.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.6
pytz==2025.2
pywin32==311
PyYAML==6.0.3
qdrant-client==1.16.2
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.10.0
tqdm==4.67.3
transformers==5.1.0
tree-sitter==0.25.2
tree-sitter-c-sharp==0.23.1
tree-sitter-embedded-template==0.25.0
tree-sitter-language-pack==0.13.0
tree-sitter-yaml==0.7.2
turbopuffer==1.15.0
typer==0.23.1
typer-slim==0.23.1
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.40.0
validators==0.35.0
vecs==0.4.5
voyageai==0.3.7
watchfiles==1.1.1
weaviate-client==4.19.2
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
